% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_auxiliary.R
\name{cov_uni_lagrangian}
\alias{cov_uni_lagrangian}
\title{Compute the univariate Lagrangian spatio-temporal stationary covariance function}
\usage{
cov_uni_lagrangian(location, scale, vx, vy)
}
\arguments{
\item{location}{An \eqn{n \times 3} matrix of coordinates.}

\item{sigma2}{A numeric constant indicating the spatial variance parameter.}

\item{scale}{A numeric constant indicating the spatial scale parameter.}

\item{nu}{A numeric constant indicating the spatial smoothness parameter.}

\item{vx}{A numeric constant indicating the x component of the transport velocity.}

\item{vy}{A numeric constant indicating the y component of the transport velocity.}

\item{nonstat}{0, 1, 2 indicating the type of nonstationarity exhibited by the
spatial field. Default is 0 (stationarity).}
}
\value{
A matrix of dimension \eqn{n \times n}.
}
\description{
\code{cov_uni_lagrangian} evaluates the univariate spatio-temporal stationary
covariance function model based on the Lagrangian transport.
}
\examples{

library(dplyr)
library(fields)

x <- seq(0, 1, length.out = 10)
y <- seq(0, 1, length.out = 10)

loc2d <- expand.grid(x, y) \%>\% as.matrix()
t <- seq(0, 10, by = 1)
loc3d <- cbind(rep(loc2d[, 1], each = length(t)), rep(loc2d[, 2], each = length(t)), t)

cov_mat <- cov_uni_lagrangian(location = loc3d, sigma2 = 1, scale = 0.5, nu = 1, 
                              vx = 0.1001, vy = 0.1001)


}
\author{
Mary Lai Salvana \email{yourlainess@gmail.com}
}
